finished dijkstra (not tested)

+++ src/MinHeapItem.java
@@ -3,6 +3,6 @@ public interface MinHeapItem {
     int getIndex();
     void setIndex(int index);
 
-    int getValue();
-    void setValue(int value);
+    double getValue();
+    void setValue(double value);
 }
